[VC] Split ScriptManager into partials and add cache
Refactors ScriptManager into multiple partial files and introduces a document cache and cached-script model to improve parsing and indexing performance. Added CachedScript and ScriptCache to track parsed scripts, dependents, content hashes, and incremental text updates; split responsibilities into ScriptManager.Editor, .Dependencies, .Indexing, and .Symbols for clearer separation of concerns; implemented dependency parsing/registration, merging and deduplication of dependency symbols, workspace indexing (partial/full modes), and symbol registry population. Also slimmed down the main ScriptManager constructor and moved file-read/parse logic into EnsureParsedAsync. These changes enable incremental updates, parallel indexing, and more maintainable code organization.
